biotech-monitor/requirements.txt
# HTTP client
httpx>=0.25.0
requests>=2.31.0

# Retry logic
tenacity>=8.2.0

# Environment variables
python-dotenv>=1.0.0

# Configuration
pyyaml>=6.0.1

# Data processing
pandas>=2.0.0
pyarrow>=14.0.0

# CLI
click>=8.1.0

# Terminal UI
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
respx>=0.20.0
responses>=0.25.0
freezegun>=1.2.0

# Type checking (optional)
mypy>=1.5.0

# Visualization
plotly>=5.18.0
kaleido>=0.2.1

# Text extraction
pymupdf>=1.23.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
